requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.3
beautifulsoup4>=4.12.3
APScheduler>=3.10.4
aiosqlite>=0.20.0
python-dotenv>=1.0.1